refactor(datacenters): converge data center functions to be an individual lib
