+++ internal/cubecos/health.go
@@ -15,6 +15,7 @@ import (
 	"github.com/bigstack-oss/bigstack-dependency-go/pkg/wait"
 	v1 "github.com/bigstack-oss/cube-cos-api/internal/definition/v1"
 	"github.com/bigstack-oss/cube-cos-api/internal/definition/v1/base"
+	"github.com/bigstack-oss/cube-cos-api/internal/definition/v1/datacenters"
 	cuberr "github.com/bigstack-oss/cube-cos-api/internal/definition/v1/errors"
 	"github.com/bigstack-oss/cube-cos-api/internal/definition/v1/event"
 	"github.com/bigstack-oss/cube-cos-api/internal/status"
@@ -39,9 +40,9 @@ var (
 )
 
 type Health struct {
-	*v1.DataCenter `json:"dataCenter,omitempty" bson:"dataCenter,omitempty"`
-	*Overall       `json:"overall,omitempty" bson:"overall"`
-	Services       []v1.Service `json:"services" bson:"services"`
+	*datacenters.DataCenter `json:"dataCenter,omitempty" bson:"dataCenter,omitempty"`
+	*Overall                `json:"overall,omitempty" bson:"overall"`
+	Services                []v1.Service `json:"services" bson:"services"`
 }
 
 type Overall struct {
